refactor(keymap): extract AnyKeycode evaluator from keycodes-utils.ts (1083→670 lines)
Move buildAnyKeycodeFunctions, tokenizer, and decodeAnyKeycode parser
into keycodes-any-keycode.ts. Re-export for backward compatibility.
